chatkit-backend/requirements.txt
openai-chatkit
openai-agents
fastapi
uvicorn[standard]
httpx
msal
python-jose[cryptography]
mcp
itsdangerous
